Minor changes to remove compilation warnings

+++ commons/tseries/timeseries.cpp
@@ -9,7 +9,7 @@ TimeSeries::TimeSeries()
 
 void TimeSeries::setThreshold(int idx, double value)
 {
-    while (mThresholds.size() <= idx)
+    while ((int)mThresholds.size() <= idx)
         mThresholds.push_back(0);
 
     mThresholds.at(idx) = value;
